+++ config/breadcrumbs/backoffice.rb
@@ -35,11 +35,31 @@ crumb :backoffice_research_area do |research_area|
 end
 
 crumb :backoffice_research_area_new do |research_area|
-  link research_area.name, new_backoffice_research_area_path(research_area)
+  link "New", new_backoffice_research_area_path(research_area)
   parent :backoffice_research_areas
 end
 
 crumb :backoffice_research_area_edit do |research_area|
-  link research_area.name, edit_backoffice_research_area_path(research_area)
-  parent :backoffice_research_areas
+  link "Edit", edit_backoffice_research_area_path(research_area)
+  parent :backoffice_research_area, research_area
+end
+
+crumb :backoffice_categories do
+  link "Categories", backoffice_categories_path
+  parent :backoffice_root
+end
+
+crumb :backoffice_category do |category|
+  link category.name, backoffice_category_path(category)
+  parent :backoffice_categories
+end
+
+crumb :backoffice_category_new do |category|
+  link "New", new_backoffice_category_path(category)
+  parent :backoffice_categories
+end
+
+crumb :backoffice_category_edit do |category|
+  link "Edit", edit_backoffice_category_path(category)
+  parent :backoffice_category, category
 end